Redirect backwards compat  🐿 v2.6.0
