+++ lib/smoke/index.js
@@ -11,7 +11,7 @@ module.exports.run = (opts) => {
 
 		process.env.SMOKE_CONFIG = configFile;
 		process.env.SMOKE_HOST = opts.host || 'http://localhost:3002';
-		process.env.SMOKE_AUTHENTICATE = opts.authenticate || false;
+		process.env.SMOKE_AUTHENTICATE = opts.authenticate;
 
 		//If running the tests, jest will already be defined on the global scope, so check it exists first.
 		if(typeof jest === 'undefined') {
