+++ lib/smoke/test.js
@@ -1,6 +1,8 @@
 /* globals , beforeAll, afterAll, test, expect */
 const puppeteer = require('puppeteer');
 
+const verifyCacheHeaders = require('./verify-cache-headers');
+
 const config = require(process.env.SMOKE_CONFIG);
 
 const host = process.env.SMOKE_HOST || 'http://localhost:3002';
@@ -17,51 +19,8 @@ const teardown = async () => {
 	await browser.close();
 };
 
-const verifyCacheHeaders = (headers, path) => {
+const REDIRECT_CODES = [301, 302, 303, 307, 308];
 
-
-	const cacheErrors = [];
-
-	if (!headers['surrogate-control'] && !headers['cache-control']) {
-		cacheErrors.push(`Each ${path} should specify a Cache-Control and/or a Surrogate-Control header`);
-	}
-	if (headers['cache-control'] && headers['cache-control'].includes('private')) {
-		if (headers['surrogate-control'] && !headers['surrogate-control'].includes('max-age=0')) {
-			cacheErrors.push(`${path} has a private cache-control, which will mean surrogate-control gets ignored by fastly`);
-		}
-	} else {
-		if (headers['surrogate-control'] && !(headers['surrogate-control'].includes('stale-while-revalidate')|| headers['surrogate-control'].includes('stale-if-error'))) {
-			cacheErrors.push(`${path} should specify stale-while-revalidate and stale-if-error cache headers`);
-		}
-
-		if (!headers['surrogate-control'] || !headers['cache-control']) {
-			cacheErrors.push(`Cachable path ${path} should specify both a Cache-Control and a Surrogate-Control header`);
-		}
-
-		if (headers['cache-control'] && headers['cache-control'].includes('public') && /max-age=[^0]/.test(headers['cache-control'])) {
-			cacheErrors.push(`${path} should not have a public Cache-Control header of max-age greater than 0`);
-		}
-
-		if (headers['surrogate-control'] && !headers['cache-control']) {
-			cacheErrors.push(`As ${path} uses surrogate-control, you should set an aoutbound cache-control header too, usually res.set('Cache-Control', res.FT_NO_CACHE)`);
-		}
-	}
-
-	if (cacheErrors.length) {
-		console.error(cacheErrors.join('\n')); // eslint-disable-line no-console
-		// eslint-disable-next-line no-console
-		console.error(`\
-n-express contains a few helpful cache constants you can use to rectify these issues:
-res.FT_NO_CACHE = 'max-age=0, no-cache, no-store, must-revalidate';
-res.FT_SHORT_CACHE = 'max-age=600, stale-while-revalidate=60, stale-if-error=86400';
-res.FT_HOUR_CACHE = 'max-age=3600, stale-while-revalidate=60, stale-if-error=86400';
-res.FT_DAY_CACHE = 'max-age=86400, stale-while-revalidate=60, stale-if-error=86400';
-res.FT_LONG_CACHE = 'max-age=86400, stale-while-revalidate=60, stale-if-error=259200';
-e.g. res.set('Cache-Control', res.FT_NO_CACHE).set('Surrogate-Control', res.FT_HOUR_CACHE);
-`);
-		throw new Error('Unwise Cache headers');
-	}
-};
 
 describe('Smoke Tests', () => {
 	beforeAll(setup);
@@ -79,34 +38,59 @@ describe('Smoke Tests', () => {
 				let response;
 				let headers;
 				let coverageReports;
+				let redirects = {};
+
+				const fullUrl = `${host}${url}`;
 
 				beforeAll(async () => {
 					if(test.cssCoverage) {
 						page.coverage.startCSSCoverage();
 					}
 
-					if(process.env.SMOKE_AUTHENTICATE) {
+					if(process.env.SMOKE_AUTHENTICATE === true && process.env.FT_NEXT_BACKEND_KEY) {
 						await page.setExtraHTTPHeaders({
 							'FT-Next-Backend-Key': process.env.FT_NEXT_BACKEND_KEY
 						});
 					}
 
-					if(config.headers) {
-						await page.setExtraHTTPHeaders(config.headers);
+					if(conf.headers) {
+						await page.setExtraHTTPHeaders(conf.headers);
 					}
 
 					if(test.headers) {
 						await page.setExtraHTTPHeaders(test.headers);
 					}
 
-					response = await page.goto(`${host}${url}`);
+					page.on('response', response => {
+						const request = response.request();
+						const status = response.status();
+						// if this response is a redirect
+						if (REDIRECT_CODES.includes(status)) {
+							redirects[request.url()] = {
+								status: status,
+								to: response.url()
+							};
+						}
+					});
+
+					response = await page.goto(fullUrl);
 					headers = response.headers();
 				});
 
 				if (test.status) {
-					it(`should return a ${test.status}`, async () => {
-						const status = await response.status();
-						expect(status).toEqual(test.status);
+					it(`should return a ${test.status}`, () => {
+
+						if(REDIRECT_CODES.includes(test.status)) {
+							const redirect = redirects[fullUrl];
+							expect(redirect.status).toEqual(test.status);
+						} else if (typeof test.status === 'string') {
+							const eventualUrl = response.url()
+							expect(eventualUrl).toEqual(`${host}${test.status}`);
+						} else {
+							const status = response.status();
+							expect(status).toEqual(test.status);
+
+						}
 					});
 				}
